Revert "fix: Make file move and DB update atomic for submissions (#1785)"
This reverts commit be6d2a402a7b3239582eb2805476c9a02e1d22d4.
