+++ app/package.json
@@ -8,10 +8,7 @@
     "debug": "nodemon --inspect ./bin/www",
     "build": "cd frontend && npm run build",
     "start": "node ./bin/www",
-    "test": "npm run test:unit",
-    "test:unit": "jest --config jest.config.unit.js --verbose --forceExit --detectOpenHandles",
-    "test:integration": "jest --config jest.config.integration.js --verbose --forceExit --detectOpenHandles",
-    "test:all": "jest --verbose --forceExit --detectOpenHandles",
+    "test": "jest --verbose --forceExit --detectOpenHandles",
     "lint": "eslint . **/www --no-fix --ext .js",
     "lint:fix": "eslint . **/www --fix --ext .js",
     "prettier": "prettier --check './**/*.js'",
